renamed compartmentalmodel.h to compartmental_model.h
