+++ cpp/memilio/compartments/compartmental_model.h
