+++ cpp/memilio/compartments/simulation.h
@@ -21,7 +21,7 @@
 #define SIMULATION_H
 
 #include "memilio/config.h"
-#include "memilio/compartments/compartmentalmodel.h"
+#include "memilio/compartments/compartmental_model.h"
 #include "memilio/utils/metaprogramming.h"
 #include "memilio/math/stepper_wrapper.h"
 #include "memilio/utils/time_series.h"
