+++ cpp/models/glct_secir/model.h
@@ -24,7 +24,7 @@
 #include "glct_secir/parameters.h"
 #include "glct_secir/infection_state.h"
 #include "memilio/epidemiology/lct_infection_state.h"
-#include "memilio/compartments/compartmentalmodel.h"
+#include "memilio/compartments/compartmental_model.h"
 #include "memilio/epidemiology/populations.h"
 #include "memilio/config.h"
 #include "memilio/utils/logging.h"
