+++ pycode/memilio-generation/memilio/generation/template/template_string.py
@@ -58,7 +58,7 @@ def includes(intermed_repr: IntermediateRepresentation) -> str:
     if intermed_repr.is_compartmentalmodel:
         substitution_string += (
             "#include \"compartments/simulation.h\"\n"
-            "#include \"compartments/compartmentalmodel.h\"\n"
+            "#include \"compartments/compartmental_model.h\"\n"
             "#include \"epidemiology/populations.h\"\n"
         )
