+++ pycode/memilio-simulation/memilio/simulation/bindings/models/osecir.cpp
@@ -22,7 +22,7 @@
 #include "pybind_util.h"
 #include "compartments/simulation.h"
 #include "compartments/flow_simulation.h"
-#include "compartments/compartmentalmodel.h"
+#include "compartments/compartmental_model.h"
 #include "epidemiology/age_group.h"
 #include "epidemiology/populations.h"
 #include "utils/custom_index_array.h"
