+++ pycode/memilio-simulation/memilio/simulation/bindings/models/oseir.cpp
@@ -25,7 +25,7 @@
 #include "utils/parameter_set.h"
 #include "compartments/simulation.h"
 #include "compartments/flow_simulation.h"
-#include "compartments/compartmentalmodel.h"
+#include "compartments/compartmental_model.h"
 #include "epidemiology/age_group.h"
 #include "epidemiology/populations.h"
